model taphonomic effects

Models taphonomy by combining the change in taphonomic conditions with the preservation potential as a function of taphonomic conditions to determine how preservation potential changes. This is then used to systematically remove (thin) the event data using thin.

Usage

apply_taphonomy(x, pres_potential, ctc)

Arguments

x

events, e.g. times/ages of fossil occurrences or their stratigraphic position.

pres_potential

function. Takes taphonomic conditions as input and returns the preservation potential (a number between 0 and 1)

ctc

function, change in taphonomic conditions (ctc) with time or stratigraphic position.

Value

numeric vector, location/timing of events (e.g. fossil occurrences) after the taphonomic filter is applied

See also

Examples

# see
#vignette("advanced_functionality")
# for details on usage